Initial working version of project users. Still implementing input tokenizer and restricting actions.

+++ paw/services/src/main/java/ar/edu/itba/services/ProjectServiceImpl.java
@@ -1,6 +1,7 @@
 package ar.edu.itba.services;
 
 import java.util.List;
+import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -10,183 +11,231 @@ import org.springframework.stereotype.Service;
 import ar.edu.itba.interfaces.ProjectDao;
 import ar.edu.itba.interfaces.ProjectService;
 import ar.edu.itba.models.Project;
+import ar.edu.itba.models.User;
 
 @Service
 public class ProjectServiceImpl implements ProjectService{
 
 	@Autowired
 	private ProjectDao projectDao;
-	
+
 	private CodeValidator codeValidator = new CodeValidator();
 
 	@Override
-	public Project createProject(final String name, final String description, final String code) {
+	public Project createProject(final User admin, final Set<User> members, final String name, final String description, final String code) {
 		if (name == null) {
 			throw new IllegalArgumentException("Project name can't be null");
 		}
-		
+
 		if (name.length() == 0) {
 			throw new IllegalArgumentException("Project name can't be empty");
 		}
-		
+
 		if (name.length() > 100) {
 			throw new IllegalArgumentException("Project name can't be longer than 100 characters");
 		}
-		
+
 		if (description == null) {
 			throw new IllegalArgumentException("Description can't be null");
 		}
-		
+
 		if (description.length() == 0) {
 			throw new IllegalArgumentException("Description can't be empty");
 		}
-		
+
 		if (description.length() > 500) {
 			throw new IllegalArgumentException("Project name can't be longer than 500 characters");
 		}
-		
+
 		if (code == null) {
 			throw new IllegalArgumentException("Project code can't be null");
 		}
-		
+
 		if (code.length() == 0) {
 			throw new IllegalArgumentException("Project code can't be empty");
 		}
-		
+
 		if (code.length() > 10) {
 			throw new IllegalArgumentException("Project code can't have more than 10 characters");
 		}
-		
+
+		if (admin == null) {
+			throw new IllegalArgumentException("Project admin can't be null");
+		}
+
+		if (members == null) {
+			throw new IllegalArgumentException("Project members can't be null");
+		}
+
 		if (!codeValidator.validate(code)) {
 			throw new IllegalArgumentException("Project code can only have numbers and lower case characters");
 		}
-		
+
 		if (projectDao.projectNameExists(name)) {
 			throw new IllegalStateException("This project name has been used already");
 		}
-		
+
 		if (projectDao.projectCodeExists(code)) {
 			throw new IllegalStateException("This project code has been used already");
 		}
 
-		return projectDao.createProject(name, description, code);
+		Project project = projectDao.createProject(admin, name, description, code);
+
+		for (User user: members) {
+			projectDao.addProjectMember(project, user);
+		}
+
+		return projectDao.getProjectById(project.projectId());
 	}
 
 	@Override
-	public void deleteProject(final Project project) {
+	public void deleteProject(final User user, final Project project) {
 		if (project == null) {
 			throw new IllegalArgumentException("Project can't be null");
 		}
-		
+
 		if (!projectDao.projectExists(project)) {
 			throw new IllegalStateException("Project doesn't exist");
 		}
-		
+
+		if (user == null) {
+			throw new IllegalArgumentException("Project admin can't be null");
+		}
+
+		if (!project.admin().equals(user)) {
+			throw new IllegalStateException("This user doesnt isn't the project admin");
+		}
+
 		projectDao.deleteProject(project);
+		projectDao.deleteProjectUser(project);
 	}
 
 	@Override
-	public Project setName(final Project project, final String name) {
+	public Project setName(final User user, final Project project, final String name) {
 		if (project == null) {
 			throw new IllegalArgumentException("Project can't be null");
 		}
-		
+
 		if (name == null) {
 			throw new IllegalArgumentException("Project name can't be null");
 		}
-		
+
 		if (name.length() == 0) {
 			throw new IllegalArgumentException("Project name can't be empty");
 		}
-		
+
 		if (name.length() > 100) {
 			throw new IllegalArgumentException("Project name can't be longer than 100 characters");
 		}
-		
+
+		if (user == null) {
+			throw new IllegalArgumentException("Project admin can't be null");
+		}
+
+		if (!project.admin().equals(user)) {
+			throw new IllegalStateException("This user doesnt isn't the project admin");
+		}
+
 		if (!projectDao.projectExists(project)) {
 			throw new IllegalStateException("Project doesn't exist");
 		}
-		
+
 		if (project.name().equals(name)) {
 			return project;
 		}
-		
+
 		if (projectDao.projectNameExists(name)) {
 			throw new IllegalStateException("This project name has been used already");
 		}		
-		
+
 		projectDao.updateName(project, name);
-		
+
 		return projectDao.getProjectById(project.projectId());
 	}
 
 	@Override
-	public Project setDescription(final Project project, final String description) {
+	public Project setDescription(final User user, final Project project, final String description) {
 		if (project == null) {
 			throw new IllegalArgumentException("Project can't be null");
 		}
-		
+
 		if (description == null) {
 			throw new IllegalArgumentException("Project description can't be null");
 		}
-		
+
 		if (description.length() == 0) {
 			throw new IllegalArgumentException("Project description can't be empty");
 		}
-		
+
 		if (description.length() > 500) {
 			throw new IllegalArgumentException("Project description can't be longer than 100 characters");
 		}
-		
+
+		if (user == null) {
+			throw new IllegalArgumentException("Project admin can't be null");
+		}
+
+		if (!project.admin().equals(user)) {
+			throw new IllegalStateException("This user doesnt isn't the project admin");
+		}
+
 		if (!projectDao.projectExists(project)) {
 			throw new IllegalStateException("Project doesn't exist");
 		}
-		
+
 		if (project.description().equals(description)) {
 			return project;
 		}		
-		
+
 		projectDao.updateDescription(project, description);
-		
+
 		return projectDao.getProjectById(project.projectId());
 	}
 
 	@Override
-	public Project setCode(final Project project, final String code) {
+	public Project setCode(final User user, final Project project, final String code) {
 		if (project == null) {
 			throw new IllegalArgumentException("Project can't be null");
 		}
-		
+
 		if (code == null) {
 			throw new IllegalArgumentException("Project code can't be null");
 		}
-		
+
 		if (code.length() == 0) {
 			throw new IllegalArgumentException("Project code can't be empty");
 		}
-		
+
 		if (code.length() > 10) {
 			throw new IllegalArgumentException("Project code can't be longer than 10 characters");
 		}
-		
+
 		if (!codeValidator.validate(code)) {
 			throw new IllegalArgumentException("Project code can only have numbers and lower case characters");
 		}
-		
+
+		if (user == null) {
+			throw new IllegalArgumentException("Project admin can't be null");
+		}
+
+		if (!project.admin().equals(user)) {
+			throw new IllegalStateException("This user doesnt isn't the project admin");
+		}
+
 		if (!projectDao.projectExists(project)) {
 			throw new IllegalStateException("Project doesn't exist");
 		}
-		
+
 		if (project.code().equals(code)) {
 			return project;
 		}
-		
+
 		if (projectDao.projectCodeExists(code)) {
 			throw new IllegalStateException("This project code has been used already");
 		}
-		
+
 		projectDao.updateCode(project, code);
-		
+
 		return projectDao.getProjectById(project.projectId());
 	}
 
@@ -195,9 +244,9 @@ public class ProjectServiceImpl implements ProjectService{
 		if (projectId < 0) {
 			throw new IllegalArgumentException("Invalid project id");
 		}
-		
+
 		Project project = projectDao.getProjectById(projectId);
-		
+
 		if (project == null) {
 			throw new IllegalStateException("Project doesn't exist");
 		} else {
@@ -206,8 +255,12 @@ public class ProjectServiceImpl implements ProjectService{
 	}
 
 	@Override
-	public List<Project> getProjects() {
-		return projectDao.getProjects();	
+	public List<Project> getProjectsForUser(final User user) {
+		if (user == null) {
+			throw new IllegalArgumentException("Can't retrieve project for null user");
+		}
+
+		return projectDao.getProjectsForUser(user);	
 	}
 
 	@Override
@@ -215,52 +268,34 @@ public class ProjectServiceImpl implements ProjectService{
 		if (code == null) {
 			throw new IllegalArgumentException("Project code can't be null");
 		}
-		
+
 		if (code.length() == 0) {
 			throw new IllegalArgumentException("Project code can't be empty");
 		}
-		
+
 		if (code.length() > 10) {
 			throw new IllegalArgumentException("Project code can't be longer than 10 characters");
 		}
-		
+
 		if (!codeValidator.validate(code)) {
 			throw new IllegalArgumentException("Project code can only have numbers and lower case characters");
 		}
-		
+
 		Project project = projectDao.getProjectByCode(code);
-		
+
 		if (project == null) {
 			throw new IllegalStateException("Project doesn't exist");
 		} else {
 			return project;
 		}
 	}	
-	
-	private class CodeValidator {
-
-    	private Pattern pattern;
-    	private Matcher matcher;
-
-    	private static final String CODE_PATTERN = "[A-Za-z0-9]+";
-
-    	public CodeValidator() {
-    		pattern = Pattern.compile(CODE_PATTERN);
-    	}
-
-    	public boolean validate(final String hex) {
-    		matcher = pattern.matcher(hex);
-    		return matcher.matches();
-    	}
-    	
-    }
 
 	@Override
 	public boolean projectCodeExists(String code) {
 		if (code == null) {
 			throw new IllegalArgumentException("Project code can't be null");
 		}
-		
+
 		return projectDao.projectCodeExists(code);
 	}
 
@@ -269,8 +304,39 @@ public class ProjectServiceImpl implements ProjectService{
 		if (name == null) {
 			throw new IllegalArgumentException("Project name can't be null");
 		}
-		
+
 		return projectDao.projectNameExists(name);
+	}
+
+	@Override
+	public List<User> getProjectMembers(Project project) {
+		if (project == null) {
+			throw new IllegalArgumentException("Project can't be null");
+		}
+
+		if (!projectDao.projectExists(project)) {
+			throw new IllegalStateException("Project doesn't exist");
+		}
+
+		return projectDao.getProjectMembers(project);
+	}
+
+	private class CodeValidator {
+
+		private Pattern pattern;
+		private Matcher matcher;
+
+		private static final String CODE_PATTERN = "[A-Za-z0-9]+";
+
+		public CodeValidator() {
+			pattern = Pattern.compile(CODE_PATTERN);
+		}
+
+		public boolean validate(final String hex) {
+			matcher = pattern.matcher(hex);
+			return matcher.matches();
+		}
+
 	}
 
 }
